moved generic definitions from coring to kering so better module dependency mapping
